task-platform/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3
email-validator>=2.1
